+++ source/pyreshaper/specification.py
@@ -6,7 +6,7 @@ the PyReshaper code is specified.  Currently all types of supported
 operations for the PyReshaper are specified with derived dypes of the
 Specification class.
 
-Copyright 2015, University Corporation for Atmospheric Research
+Copyright 2016, University Corporation for Atmospheric Research
 See the LICENSE.rst file for details
 """
 
@@ -51,6 +51,7 @@ class Specifier(object):
                  prefix='tseries.',
                  suffix='.nc',
                  metadata=[],
+                 backend='netCDF4',
                  **kwargs):
         """
         Initializes the internal data with optional arguments.
@@ -99,6 +100,9 @@ class Specifier(object):
         # List of time-variant variables that should be included in all
         #  output files.
         self.time_variant_metadata = metadata
+
+        # Store the netCDF I/O backend name
+        self.io_backend = backend
 
         # Optional arguments associated with the reshaper operation
         self.options = kwargs
@@ -164,6 +168,11 @@ class Specifier(object):
                           "given as strings"
                 raise TypeError(err_msg)
 
+        # Validate the type of the backend
+        if not isinstance(self.io_backend, str):
+            err_msg = "I/O backend must be given as a string"
+            raise TypeError(err_msg)
+
     def validate_values(self):
         """
         Method to validate the values of the Specifier data.
@@ -173,7 +182,7 @@ class Specifier(object):
         We impose the (somewhat arbitrary) rule that the Specifier
         should not validate values what require "cracking" open the
         input files themselves.  Hence, we validate values that can
-        be checked without any PyNIO file I/O (including reading the
+        be checked without any NetCDF file I/O (including reading the
         header information).
 
         This method will correct some input if it is safe to do so.
@@ -223,6 +232,8 @@ class Specifier(object):
         # Validate the output file suffix string (should end in .nc)
         if (self.output_file_suffix[-3:] != '.nc'):
             self.output_file_suffix += '.nc'
+
+        # Backend validated when PyReshaper is run ONLY!
 
     def write(self, fname):
         """