+++ source/pyreshaper/test/reshaperTests.py
@@ -1,7 +1,7 @@
 """
 Parallel Tests for the Reshaper class
 
-Copyright 2015, University Corporation for Atmospheric Research
+Copyright 2016, University Corporation for Atmospheric Research
 See the LICENSE.rst file for details
 """
 
@@ -14,7 +14,7 @@ from os import linesep as eol
 from os import remove
 from mpi4py import MPI
 
-from pyreshaper.reshaper import Slice2SeriesReshaper, create_reshaper
+from pyreshaper.reshaper import Reshaper, create_reshaper
 from pyreshaper.specification import Specifier
 
 import mkTestData
@@ -22,7 +22,7 @@ import mkTestData
 MPI_COMM_WORLD = MPI.COMM_WORLD
 
 
-class S2SReshaperTests(unittest.TestCase):
+class ReshaperTests(unittest.TestCase):
 
     def setUp(self):
 
@@ -115,7 +115,7 @@ class S2SReshaperTests(unittest.TestCase):
             rshpr = create_reshaper(spec, serial=serial, verbosity=verbosity,
                                     wmode=wmode)
             self._assertion("type(reshaper)", type(rshpr),
-                            Slice2SeriesReshaper)
+                            Reshaper)
 
     def test_create_reshaper_serial_V0_W(self):
         self._test_create_reshaper(serial=True, verbosity=0, wmode='w')
@@ -332,7 +332,7 @@ if __name__ == "__main__":
     MPI_COMM_WORLD.Barrier()
 
     mystream = StringIO()
-    tests = unittest.TestLoader().loadTestsFromTestCase(S2SReshaperTests)
+    tests = unittest.TestLoader().loadTestsFromTestCase(ReshaperTests)
     unittest.TextTestRunner(stream=mystream).run(tests)
     MPI_COMM_WORLD.Barrier()